config: add configuration loading
Configuration is ensured to exists every time the application is ran.
If no configuration `config` file or `.gong` config directory is found,
those will be created automatically by the system.

Add new package `fs`, which stands for filesystem, strictly for
filesystem related behaviour.

fs: add `EnsureDir` function

EnsureDir checks if a given directory exists within the path and creates it if it does not.
If the given pathname is a file and not a directory the funtion returns an error.